databricks-app/requirements.txt
folium
streamlit_folium
databricks-sdk
plotly